v0.2
Clean up logging support:
 * Default console logging to level WARNING
 * Log all vprint messages to the log file (if specified)
 * Only log INFO and DEBUG messages to the log file (if specified)

Clean up script structure and execution flow to enable the contents of
the script to be eecuted or imported.  Nest execution flow, environment
setup, parsing and etc. underneath the ``if __name__ == "__main__”``
statement.  Minimize the use of global variables, and pass data as
parameters to the functions.